% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{check_dep_word}
\alias{check_dep_word}
\title{Check and convert dependency word(s)}
\usage{
check_dep_word(x)
}
\arguments{
\item{x}{A character vector of dependency words}
}
\value{
A character vector of modified dependency words
}
\description{
Check and convert dependency word(s)
}
\keyword{internal}
