% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{conditional_change}
\alias{conditional_change}
\title{Conditionally change a string}
\usage{
conditional_change(x, from, to)
}
\arguments{
\item{x}{A character vector}

\item{from}{A character vector of words to change from}

\item{to}{A string to change to}
}
\value{
A string
}
\description{
Conditionally change a string
}
\keyword{internal}
