% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dupp}
\alias{dupp}
\title{Probability mass function of discrete power law greater than or equal to u}
\usage{
dupp(x, u, alpha)
}
\arguments{
\item{x}{data vector of positive integers}

\item{u}{scalar, non-negative integer threshold}

\item{alpha}{scalar, a numeric value greater than 1.0}
}
\value{
A numeric vector of the same length as x
}
\description{
Probability mass function of discrete power law greater than or equal to u
}
