% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_dep_all}
\alias{get_dep_all}
\title{Obtain one type of dependencies of a package directly. This is essentially the same as chaining get}
\usage{
get_dep_all(name, type)
}
\arguments{
\item{name}{String, name of the package}

\item{type}{One of the following dependency words: "Depends", "Imports", "LinkingTo", "Suggests", "Reverse_depends", "Reverse_imports", "Reverse_linking_to", "Reverse_suggests"}
}
\value{
A string vector of dependencies
}
\description{
Obtain one type of dependencies of a package directly. This is essentially the same as chaining get
}
\examples{
get_dep_all("dplyr", "Imports")
get_dep_all("MASS", "Depends")
}
