% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_dep_all_packages}
\alias{get_dep_all_packages}
\title{Obtain dependencies of all CRAN packages}
\usage{
get_dep_all_packages()
}
\value{
A data frame of dependencies of all CRAN packages
}
\description{
Obtain dependencies of all CRAN packages
}
\examples{
df.cran <- crandep:::get_dep_all_packages()
}
\keyword{internal}
