% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_dep_str}
\alias{get_dep_str}
\title{Find string corresponding to "Imports", "Depends" etc.}
\usage{
get_dep_str(v, x)
}
\arguments{
\item{v}{A vector of strings}

\item{x}{One of the following dependency words: "Depends", "Imports", "LinkingTo", "Suggests", "Reverse_depends", "Reverse_imports", "Reverse_linking_to", "Reverse_suggests"}
}
\value{
A scalar, concatenated string of dependencies
}
\description{
Find string corresponding to "Imports", "Depends" etc.
}
