% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{html_text_vec}
\alias{html_text_vec}
\title{Scrap the page (of a package) as a text vector}
\usage{
html_text_vec(url)
}
\arguments{
\item{url}{An URL}
}
\value{
A string vector of the html text of the page according to the url
}
\description{
Scrap the page (of a package) as a text vector
}
\keyword{internal}
