% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{list_of_dep}
\alias{list_of_dep}
\title{Split a string to a list of dependencies}
\usage{
list_of_dep(x)
}
\arguments{
\item{x}{A scalar string, preferably an output of get_dep_str()}
}
\value{
A string vector of dependencies
}
\description{
Split a string to a list of dependencies
}
