% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{lpost_bulk_wrapper}
\alias{lpost_bulk_wrapper}
\title{Wrapper of lpost_bulk, assuming power law (theta = 1.0)}
\usage{
lpost_bulk_wrapper(alpha, ...)
}
\arguments{
\item{alpha}{A scalar, positive}

\item{...}{Other arguments passed to lpost_bulk}
}
\value{
A scalar of the log-posterior density
}
\description{
Wrapper of lpost_bulk, assuming power law (theta = 1.0)
}
\keyword{internal}
