% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{lpost_mix2_constrained}
\alias{lpost_mix2_constrained}
\title{Wrapper of lpost_mix2, assuming power law (theta = 1.0) & contrained (alpha > 1.0, xi < 1.0 / (alpha - 1.0))}
\usage{
lpost_mix2_constrained(par, ...)
}
\arguments{
\item{par}{parameter vector of length 3, with elements alpha, shape and sigma}

\item{...}{Other arguments passed to lpost_mix2}
}
\value{
A scalar of the log-posterior density
}
\description{
Wrapper of lpost_mix2, assuming power law (theta = 1.0) & contrained (alpha > 1.0, xi < 1.0 / (alpha - 1.0))
}
\keyword{internal}
