% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{lpost_pol_wrapper}
\alias{lpost_pol_wrapper}
\title{Wrapper of lpost_pol, assuming power law (theta = 1.0)}
\usage{
lpost_pol_wrapper(alpha, x, count, ...)
}
\arguments{
\item{alpha}{A scalar, positive}

\item{...}{Other arguments passed to lpost_pol}
}
\value{
A scalar of the log-posterior density
}
\description{
Wrapper of lpost_pol, assuming power law (theta = 1.0)
}
\keyword{internal}
