% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{lpost_pow}
\alias{lpost_pow}
\title{Unnormalised log-posterior density of discrete power law}
\usage{
lpost_pow(alpha, df, m_alpha, s_alpha)
}
\arguments{
\item{alpha}{Real number greater than 1}

\item{df}{A data frame with at least two columns, x & count}

\item{m_alpha}{Real number, mean of the prior normal distribution for alpha}

\item{s_alpha}{Positive real number, standard deviation of the prior normal distribution for alpha}
}
\value{
A real number
}
\description{
Unnormalised log-posterior density of discrete power law
}
\keyword{internal}
