% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep.R
\name{reshape_dep}
\alias{reshape_dep}
\title{Reshape the data frame of dependencies}
\usage{
reshape_dep(x, names)
}
\arguments{
\item{x}{A character vector of dependencies, each element of which corresponds to an individual package}

\item{names}{A character vector of package names of the same length as x}
}
\value{
A data frame of dependencies
}
\description{
Reshape the data frame of dependencies
}
\keyword{internal}
