% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{unnest_dep}
\alias{unnest_dep}
\title{Obtain the data frame of one kind of dependencies}
\usage{
unnest_dep(df, type)
}
\arguments{
\item{df}{A data frame with at least two columns: name - a string vector, and <type> - a list of string vectors}

\item{type}{The dependency type desired}
}
\value{
A data frame with two columns: from & to
}
\description{
Obtain the data frame of one kind of dependencies
}
