% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{df_to_graph}
\alias{df_to_graph}
\title{Construct the giant component of the network from two data frames}
\usage{
df_to_graph(edgelist, nodelist)
}
\arguments{
\item{edgelist}{A data frame with (at least) two columns: from and to, preferably an output of unnest_dep()}

\item{nodelist}{A data frame with (at least) one column: name, that contains the nodes to include}
}
\value{
An igraph object & a connected graph
}
\description{
Construct the giant component of the network from two data frames
}
