% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep.R
\name{get_dep_all_packages}
\alias{get_dep_all_packages}
\title{Dependencies of all CRAN packages}
\usage{
get_dep_all_packages()
}
\value{
A data frame of dependencies of all CRAN packages
}
\description{
\code{get_dep_all_packages} returns the data frame of dependencies of all packages currently available on CRAN.
}
\details{
Unlike \code{get_dep} and \code{get_dep_df}, there is no boolean argument `scrape`, as it is much faster to obtain the dependencies of all packages via `tools::CRAN_package_db()`.
}
\examples{
df.cran <- get_dep_all_packages()
}
