% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep.R
\name{get_dep_all_packages}
\alias{get_dep_all_packages}
\title{Dependencies of all CRAN packages}
\usage{
get_dep_all_packages()
}
\value{
A list of two data frames, one the names of all CRAN packages, the other their dependencies
}
\description{
\code{get_dep_all_packages} returns the data frame of dependencies of all packages currently available on CRAN.
}
\examples{
\dontrun{
df.cran <- get_dep_all_packages()
}
}
\seealso{
\code{\link{get_dep}} for multiple types of dependencies, and \code{\link{get_graph_all_packages}} for obtaining directly a network of dependencies as an igraph object
}
