% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_dep_vec}
\alias{get_dep_vec}
\title{Split a string to a list of dependencies}
\usage{
get_dep_vec(x)
}
\arguments{
\item{x}{A scalar string, preferably an output of get_dep_str()}
}
\value{
A string vector of dependencies
}
\description{
Split a string to a list of dependencies
}
