% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Supp}
\alias{Supp}
\title{Survival function of discrete power law greater than or equal to u}
\usage{
Supp(x, u, alpha)
}
\arguments{
\item{x}{data vector of positive integers}

\item{u}{scalar, non-negative integer threshold}

\item{alpha}{scalar, a numeric value greater than 1.0}
}
\value{
A numeric vector of the same length as x
}
\description{
Survival function of discrete power law greater than or equal to u
}
