% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{unnest_dep}
\alias{unnest_dep}
\title{Obtain the data frame of one kind of dependencies}
\usage{
unnest_dep(df, type)
}
\arguments{
\item{df}{A data frame with at least two columns: name - a string vector, and <type> - a list of string vectors}

\item{type}{The dependency type desired}
}
\value{
A data frame with two columns: from & to
}
\description{
Obtain the data frame of one kind of dependencies
}
\examples{
if (requireNamespace("tibble", quietly = TRUE)) {
    name.vec <- c("dplyr", "MASS", "Rcpp")
    depends.vec <- rep(NA, length(name.vec))
    imports.vec <- rep(NA, length(name.vec))
    for (i in seq_along(name.vec)) {
        html0 <- html_text_vec(cran_url(name.vec[i]))
        depends.vec[i] <- list(get_dep_vec(get_dep_str(html0, "Depends")))
        imports.vec[i] <- list(get_dep_vec(get_dep_str(html0, "Imports")))
    }
    df0 <- tibble::tibble(name = name.vec, depends = depends.vec, imports = imports.vec)
    unnest_dep(df0, depends)
    unnest_dep(df0, imports)
}
}
